Move effected policies component to common management components
