+++ x-pack/plugins/security_solution/public/management/components/effected_policy_select/effected_policy_select.tsx
@@ -23,11 +23,11 @@ import { i18n } from '@kbn/i18n';
 import { EuiSelectableOption } from '@elastic/eui/src/components/selectable/selectable_option';
 import { FormattedMessage } from '@kbn/i18n/react';
 import styled from 'styled-components';
-import { PolicyData } from '../../../../../../../common/endpoint/types';
-import { getPolicyDetailPath } from '../../../../../common/routing';
-import { useAppUrl } from '../../../../../../common/lib/kibana/hooks';
-import { LinkToApp } from '../../../../../../common/components/endpoint/link_to_app';
-import { useTestIdGenerator } from '../../../../../components/hooks/use_test_id_generator';
+import { PolicyData } from '../../../../common/endpoint/types';
+import { LinkToApp } from '../../../common/components/endpoint/link_to_app';
+import { getPolicyDetailPath } from '../../common/routing';
+import { useTestIdGenerator } from '../hooks/use_test_id_generator';
+import { useAppUrl } from '../../../common/lib/kibana/hooks';
 
 const NOOP = () => {};
 const DEFAULT_LIST_PROPS: EuiSelectableProps['listProps'] = { bordered: true, showIcons: false };
