+++ x-pack/plugins/security_solution/public/management/components/effected_policy_select/index.ts
